Refactor: libcrmcommon: separate ACL code into its own source file
Before:
  1301 utils.c
  5219 xml.c

After:
   721 acl.c
  1204 utils.c
  4619 xml.c

This contains no substantive code changes, but has some refactoring to
enable the split. This adds a crmcommon_private.h header for the sole
use of the library.
