+++ lib/common/utils.c
@@ -1037,103 +1037,6 @@ stonith_ipc_server_init(qb_ipcs_service_t **ipcs, struct qb_ipcs_service_handler
     }
 }
 
-bool
-pcmk_acl_required(const char *user) 
-{
-#if ENABLE_ACL
-    if(user == NULL || strlen(user) == 0) {
-        crm_trace("no user set");
-        return FALSE;
-
-    } else if (strcmp(user, CRM_DAEMON_USER) == 0) {
-        return FALSE;
-
-    } else if (strcmp(user, "root") == 0) {
-        return FALSE;
-    }
-    crm_trace("acls required for %s", user);
-    return TRUE;
-#else
-    crm_trace("acls not supported");
-    return FALSE;
-#endif
-}
-
-#if ENABLE_ACL
-char *
-uid2username(uid_t uid)
-{
-    struct passwd *pwent = getpwuid(uid);
-
-    if (pwent == NULL) {
-        crm_perror(LOG_INFO, "Cannot get user details for user ID %d", uid);
-        return NULL;
-    }
-    return strdup(pwent->pw_name);
-}
-
-const char *
-crm_acl_get_set_user(xmlNode * request, const char *field, const char *peer_user)
-{
-    static const char *effective_user = NULL;
-    const char *requested_user = NULL;
-    const char *user = NULL;
-
-    if(effective_user == NULL) {
-        effective_user = uid2username(geteuid());
-        if (effective_user == NULL) {
-            return NULL;
-        }
-    }
-
-    requested_user = crm_element_value(request, XML_ACL_TAG_USER);
-    if(requested_user == NULL) {
-        /* @COMPAT rolling upgrades <=1.1.11
-         *
-         * field is checked for backward compatibility with older versions that
-         * did not use XML_ACL_TAG_USER.
-         */
-        requested_user = crm_element_value(request, field);
-    }
-
-    if (is_privileged(effective_user) == FALSE) {
-        /* We're not running as a privileged user, set or overwrite any existing value for $XML_ACL_TAG_USER */
-        user = effective_user;
-
-    } else if(peer_user == NULL && requested_user == NULL) {
-        /* No user known or requested, use 'effective_user' and make sure one is set for the request */
-        user = effective_user;
-
-    } else if(peer_user == NULL) {
-        /* No user known, trusting 'requested_user' */
-        user = requested_user;
-
-    } else if (is_privileged(peer_user) == FALSE) {
-        /* The peer is not a privileged user, set or overwrite any existing value for $XML_ACL_TAG_USER */
-        user = peer_user;
-
-    } else if (requested_user == NULL) {
-        /* Even if we're privileged, make sure there is always a value set */
-        user = peer_user;
-
-    } else {
-        /* Legal delegation to 'requested_user' */
-        user = requested_user;
-    }
-
-    // This requires pointer comparison, not string comparison
-    if(user != crm_element_value(request, XML_ACL_TAG_USER)) {
-        crm_xml_add(request, XML_ACL_TAG_USER, user);
-    }
-
-    if(field != NULL && user != crm_element_value(request, field)) {
-        crm_xml_add(request, field, user);
-    }
-
-    return requested_user;
-}
-#endif
-
 void *
 find_library_function(void **handle, const char *lib, const char *fn, gboolean fatal)
 {
